Move templated inference of level stream into utilities
